requirements.txt
requests
socket
whois
re
subprocess
ssl
datetime
ipaddress
dns.resolver
random